apps/smart-meal-planner-backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
boto3==1.34.93
botocore==1.34.93
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
frozenlist==1.5.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jmespath==1.0.1
multidict==6.1.0
openai==0.28.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
python-json-logger==2.0.7
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
stripe==8.9.0
paypalrestsdk==1.13.3  # PayPal payment processing SDK - updated
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3